feat: implement Modules: rewrite to ESM notation
